requirements-ocr.txt
pillow>=10.4.0
pytesseract>=0.3.13
pdf2image>=1.17.0
PyMuPDF>=1.24.0
